requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0

# Deep learning
torch>=1.9.0

# Optional but recommended for improved embeddings
sentence-transformers>=2.2.0

# Optional for better theme extraction
scikit-learn>=1.0.0

# Vector database (choose one)
faiss-cpu>=1.7.0
# faiss-gpu>=1.7.0  # Use this if you have GPU

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# For quantum computing (optional - for future enhancements)
# qiskit>=0.39.0
